Add iter() method to Deb822LikeParagraph trait
This adds an iter() method to the Deb822LikeParagraph trait that
returns an iterator over all key-value pairs in a paragraph. This will
be used as a building block for implementing pattern matching support
for HashMap fields in the derive macros (issue #325).

The implementation returns `impl Iterator` for better performance
compared to boxed trait objects.